Removed unused enums. Changed conversion functions to C++17 if-constexpr. Added Certificate Parsing code for iOS using only Chromium code.
