+++ base/mac/BUILD.gn
@@ -1,24 +1,14 @@
-# Copyright (c) 2020 The Brave Authors. All rights reserved.
+# Copyright (c) 2022 The Brave Authors. All rights reserved.
 # This Source Code Form is subject to the terms of the Mozilla Public
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
-import("//build/config/ios/rules.gni")
-import("//ios/build/config.gni")
+assert(is_apple)
 
-source_set("common") {
+source_set("mac") {
   configs += [ "//build/config/compiler:enable_arc" ]
 
-  sources = [
-    "common_operations.h",
-    "common_operations.mm",
-  ]
-
-  deps = [
-    "//base",
-    "//net",
-    "//url",
-  ]
+  sources = [ "conversions.h" ]
 
   frameworks = [ "Foundation.framework" ]
 }
