+++ ios/browser/api/certificate/models/BUILD.gn
@@ -3,13 +3,8 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
-import("//build/config/ios/rules.gni")
-import("//ios/build/config.gni")
-
 source_set("models") {
   configs += [ "//build/config/compiler:enable_arc" ]
-
-  defines = [ "BRAVE_CORE_IOS_INCLUDE" ]
 
   sources = [
     "brave_certificate_enums.h",
@@ -26,6 +21,7 @@ source_set("models") {
 
   deps = [
     "//base",
+    "//brave/base/mac",
     "//brave/ios/browser/api/certificate/utils",
     "//brave/ios/browser/api/common",
     "//crypto",
