+++ ios/browser/api/certificate/models/brave_certificate_rdns_sequence.mm
@@ -5,9 +5,9 @@
 
 #include "brave/ios/browser/api/certificate/models/brave_certificate_rdns_sequence.h"
 #include "base/strings/sys_string_conversions.h"
+#include "brave/base/mac/conversions.h"
 #include "brave/ios/browser/api/certificate/utils/brave_certificate_utils.h"
 #include "brave/ios/browser/api/certificate/utils/brave_certificate_x509_utils.h"
-#include "brave/ios/browser/api/common/conversions.h"
 #include "net/cert/x509_cert_types.h"
 #include "net/der/input.h"
