+++ ios/browser/api/certificate/utils/BUILD.gn
@@ -3,9 +3,6 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
-import("//build/config/ios/rules.gni")
-import("//ios/build/config.gni")
-
 source_set("utils") {
   configs += [ "//build/config/compiler:enable_arc" ]
