+++ ios/browser/api/certificate/utils/brave_certificate_utils.mm
@@ -4,7 +4,6 @@
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 #include "brave/ios/browser/api/certificate/utils/brave_certificate_utils.h"
-#include "base/strings/string_number_conversions.h"
 #include "base/strings/sys_string_conversions.h"
 
 #if !defined(__has_feature) || !__has_feature(objc_arc)
@@ -13,12 +12,8 @@
 
 namespace certificate {
 namespace utils {
-NSData* NSDataFromString(const std::string& str) {
-  return str.empty()
-             ? [[NSData alloc] init]
-             : [NSData dataWithBytes:reinterpret_cast<const std::uint8_t*>(
-                                         str.c_str())
-                              length:str.size()];
+NSData* NSStringToData(const std::string& str) {
+  return [base::SysUTF8ToNSString(str) dataUsingEncoding:NSUTF8StringEncoding];
 }
 }  // namespace utils
 }  // namespace certificate
